+++ nls/root/strings.js
@@ -4,13 +4,20 @@
 
 define({
     BUTTON_CANCEL:          "Cancel",
+    BUTTON_CLOSE:           "Close",
     BUTTON_COMMIT:          "Commit",
     BUTTON_OK:              "OK",
     BUTTON_RESET:           "Reset index",
-    
+    BUTTON_SAVE:            "Save",
+    CHANGELOG:              "Changelog",
     ENTER_COMMIT_MESSAGE:   "Enter commit message",
-    
+    FOR_WINDOWS_USERS:      "For Windows users",
+    GIT_IS_IN_PATH:         "Git can be called from anywhere (is in system path)",
+    GIT_SETTINGS:           "Git Settings\u2026",
     NOTHING_TO_COMMIT:      "Nothing to commit, working directory clean.",
-    
+    PATH_TO_GIT_EXECUTABLE: "Path to Git executable",
+    PATH_TO_GIT_MSYSGIT:    "Path to msysgit folder",
+    Q_RESTART_BRACKETS:     "Do you wish to restart Brackets to apply new settings?",
+    RESTART:                "Restart",
     USING_GIT_VERSION:      "Using Git version"
 });