+++ src/ChangelogDialog.js
@@ -1,0 +1,29 @@
+/*jslint plusplus: true, vars: true, nomen: true */
+/*global $, brackets, define, Mustache */
+
+define(function (require, exports) {
+    "use strict";
+    
+    var Dialogs                    = brackets.getModule("widgets/Dialogs"),
+        FileEntry                  = brackets.getModule("file/NativeFileSystem").NativeFileSystem.FileEntry,
+        FileUtils                  = brackets.getModule("file/FileUtils"),
+        Strings                    = require("../strings"),
+        changelogDialogTemplate    = require("text!htmlContent/changelog-dialog.html"),
+        compiledTemplate = Mustache.render(changelogDialogTemplate, Strings);
+    
+    var dialog,
+        preferences;
+    
+    exports.show = function (prefs) {
+        if (prefs) {
+            preferences = prefs;
+        }
+        
+        dialog = Dialogs.showModalDialogUsingTemplate(compiledTemplate);
+        
+        FileUtils.readAsText(new FileEntry(preferences.getValue("extensionDirectory") + "CHANGELOG.md")).done(function (content) {
+            $("#git-changelog", dialog.getElement()).text(content);
+        });
+    };
+    
+});